Updated datapath_tb.vhd
Realised that datapath does seem to work better than I realised last time. The issue was that I previously did not define a value for MD_select in the testbench so the register was never given a defined input from MuxD. Upon closer inspection of other testbenches the oVerflow flag also seems to work as intended.