Completed datapath_tb
Successfully completed the testbench for datapath.vhd by showing that variants of all input signals produce the required output signals. This shows that the port maps within the datapath are defined correctly, which is all that is needed as the logic of the components is already proven to work in previous testbenches.