Completed register file, 4 to 16 decoder & their testbenches
Implemented decoder_4to16 and its testbench to implement the decoding for R8.
Implemented all logic for register file with added r8 including additional signals, multiplexers and the decoder. If the MSB of any address vector is set, then the corresponding input/output will always be connected to r8 no matter what the 3 LSBs of that select vector is. Also fixed an error in port maps where the MSB in address for r1 to r7 was mapped to the LSB and vice versa.